+++ projects/ngx-bootstrap-icons-lib/src/lib/icons/bookmark-fill.ts
@@ -1,10 +1,10 @@
 /**
- * BookmarkFill.
+ * bookmarkFill.
  *
  * @link https://icons.getbootstrap.com/icons/bookmark-fill/
  */
 
 /** eslint-disable max-len */
-export const BookmarkFill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-bookmark-fill" viewBox="0 0 16 16">
+export const bookmarkFill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-bookmark-fill" viewBox="0 0 16 16">
   <path fill-rule="evenodd" d="M2 2a2 2 0 0 1 2-2h8a2 2 0 0 1 2 2v13.5a.5.5 0 0 1-.74.439L8 13.069l-5.26 2.87A.5.5 0 0 1 2 15.5V2z"/>
 </svg>`;