+++ projects/ngx-bootstrap-icons-lib/src/lib/icons/calendar3-fill.ts
@@ -1,11 +1,11 @@
 /**
- * Calendar3Fill.
+ * calendar3Fill.
  *
  * @link https://icons.getbootstrap.com/icons/calendar3-fill/
  */
 
 /** eslint-disable max-len */
-export const Calendar3Fill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-calendar3-fill" viewBox="0 0 16 16">
+export const calendar3Fill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-calendar3-fill" viewBox="0 0 16 16">
   <path d="M0 2a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2H0z"/>
   <path fill-rule="evenodd" d="M0 3h16v11a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V3z"/>
 </svg>`;