+++ projects/ngx-bootstrap-icons-lib/src/lib/icons/calendar3-range-fill.ts
@@ -1,10 +1,10 @@
 /**
- * Calendar3RangeFill.
+ * calendar3RangeFill.
  *
  * @link https://icons.getbootstrap.com/icons/calendar3-range-fill/
  */
 
 /** eslint-disable max-len */
-export const Calendar3RangeFill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-calendar3-range-fill" viewBox="0 0 16 16">
+export const calendar3RangeFill = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-calendar3-range-fill" viewBox="0 0 16 16">
   <path fill-rule="evenodd" d="M2 0a2 2 0 0 0-2 2h16a2 2 0 0 0-2-2H2zm14 3H0v5h6a1 1 0 0 1 0 2H0v4a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7h-6a1 1 0 1 1 0-2h6V3z"/>
 </svg>`;