+++ projects/ngx-bootstrap-icons-lib/src/lib/icons/toggle-off.ts
@@ -1,10 +1,10 @@
 /**
- * ToggleOff.
+ * toggleOff.
  *
  * @link https://icons.getbootstrap.com/icons/toggle-off/
  */
 
 /** eslint-disable max-len */
-export const ToggleOff = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-toggle-off" viewBox="0 0 16 16">
+export const toggleOff = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-toggle-off" viewBox="0 0 16 16">
   <path fill-rule="evenodd" d="M11 4a4 4 0 0 1 0 8H8a4.992 4.992 0 0 0 2-4 4.992 4.992 0 0 0-2-4h3zm-6 8a4 4 0 1 1 0-8 4 4 0 0 1 0 8zM0 8a5 5 0 0 0 5 5h6a5 5 0 0 0 0-10H5a5 5 0 0 0-5 5z"/>
 </svg>`;