+++ projects/ngx-bootstrap-icons-lib/src/lib/icons/union.ts
@@ -1,10 +1,10 @@
 /**
- * Union.
+ * union.
  *
  * @link https://icons.getbootstrap.com/icons/union/
  */
 
 /** eslint-disable max-len */
-export const Union = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-union" viewBox="0 0 16 16">
+export const union = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-union" viewBox="0 0 16 16">
   <path fill-rule="evenodd" d="M0 2a2 2 0 0 1 2-2h8a2 2 0 0 1 2 2v2h2a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2v-2H2a2 2 0 0 1-2-2V2z"/>
 </svg>`;