+++ projects/ngx-bootstrap-icons-lib/src/lib/providers/icon.provider.ts
@@ -1,3 +1,3 @@
 export class Icons {
-  constructor(private icons: object) { }
+  constructor(private _icons: object) { }
 }