+++ src/app/app.module.ts
@@ -6,10 +6,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgxBootstrapIconsModule } from 'projects/ngx-bootstrap-icons-lib/src/lib/ngx-bootstrap-icons.module';
 
 import { ClipboardModule } from 'ngx-clipboard';
+import { allIcons } from 'projects/ngx-bootstrap-icons-lib/src/lib/icons/all';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
-import { allIcons } from 'projects/ngx-bootstrap-icons-lib/src/lib/icons/all';
 
 registerLocaleData(ru);